replace progress bar with wave-form
